updates due to jpa 2.1

+++ canvas-dao/canvas-dao-api/src/main/java/org/renci/hearsay/canvas/var/dao/model/AssemblyLocation.java
@@ -34,7 +34,7 @@ public class AssemblyLocation implements Persistable {
     private Boolean homozygous;
 
     @Column(name = "genotype_qual")
-    private Float genotypeQual;
+    private Double genotypeQual;
 
     public AssemblyLocation() {
         super();
@@ -72,11 +72,11 @@ public class AssemblyLocation implements Persistable {
         this.homozygous = homozygous;
     }
 
-    public Float getGenotypeQual() {
+    public Double getGenotypeQual() {
         return genotypeQual;
     }
 
-    public void setGenotypeQual(Float genotypeQual) {
+    public void setGenotypeQual(Double genotypeQual) {
         this.genotypeQual = genotypeQual;
     }
