+++ canvas-dao/canvas-dao-api/src/main/java/org/renci/hearsay/canvas/var/dao/model/AssemblyLocationVariantQC.java
@@ -33,19 +33,19 @@ public class AssemblyLocationVariantQC implements Persistable {
     private Integer depth;
 
     @Column(name = "qd")
-    private Float qd;
+    private Double qd;
 
     @Column(name = "read_pos_rank_sum")
-    private Float readPosRankSum;
+    private Double readPosRankSum;
 
     @Column(name = "frac_reads_with_dels")
-    private Float fracReadsWithDels;
+    private Double fracReadsWithDels;
 
     @Column(name = "hrun")
     private Integer hrun;
 
     @Column(name = "strand_score")
-    private Float strandScore;
+    private Double strandScore;
 
     @Column(name = "ref_depth")
     private Integer refDepth;
@@ -89,27 +89,27 @@ public class AssemblyLocationVariantQC implements Persistable {
         this.depth = depth;
     }
 
-    public Float getQd() {
+    public Double getQd() {
         return qd;
     }
 
-    public void setQd(Float qd) {
+    public void setQd(Double qd) {
         this.qd = qd;
     }
 
-    public Float getReadPosRankSum() {
+    public Double getReadPosRankSum() {
         return readPosRankSum;
     }
 
-    public void setReadPosRankSum(Float readPosRankSum) {
+    public void setReadPosRankSum(Double readPosRankSum) {
         this.readPosRankSum = readPosRankSum;
     }
 
-    public Float getFracReadsWithDels() {
+    public Double getFracReadsWithDels() {
         return fracReadsWithDels;
     }
 
-    public void setFracReadsWithDels(Float fracReadsWithDels) {
+    public void setFracReadsWithDels(Double fracReadsWithDels) {
         this.fracReadsWithDels = fracReadsWithDels;
     }
 
@@ -121,11 +121,11 @@ public class AssemblyLocationVariantQC implements Persistable {
         this.hrun = hrun;
     }
 
-    public Float getStrandScore() {
+    public Double getStrandScore() {
         return strandScore;
     }
 
-    public void setStrandScore(Float strandScore) {
+    public void setStrandScore(Double strandScore) {
         this.strandScore = strandScore;
     }
